refactor(infra): move pkg/{cache,cicd,compute,database,dns,loadbalancer,network,observability,secrets,storage,streaming,waf} under pkg/aws/
Phase 0 (path-only): no behavioral change. All 12 module directories
relocated under pkg/aws/ via git mv (history preserved). Import paths
updated in main.go, all module sources, and tests under test/ and
pkg/aws/*/. pkg/config/ stays at the root as a shared (cloud-agnostic)
loader.

just test-infra still passes: cicd, compute, loadbalancer, observability,
streaming, waf, config, test all green.

Refs #477.
