+++ infra/main.go
@@ -6,19 +6,19 @@ import (
 	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
 	"github.com/pulumi/pulumi/sdk/v3/go/pulumi/config"
 
-	"github.com/kaizen-experimentation/infra/pkg/cache"
-	"github.com/kaizen-experimentation/infra/pkg/cicd"
-	"github.com/kaizen-experimentation/infra/pkg/compute"
+	"github.com/kaizen-experimentation/infra/pkg/aws/cache"
+	"github.com/kaizen-experimentation/infra/pkg/aws/cicd"
+	"github.com/kaizen-experimentation/infra/pkg/aws/compute"
 	kconfig "github.com/kaizen-experimentation/infra/pkg/config"
-	"github.com/kaizen-experimentation/infra/pkg/database"
-	"github.com/kaizen-experimentation/infra/pkg/dns"
-	"github.com/kaizen-experimentation/infra/pkg/loadbalancer"
-	"github.com/kaizen-experimentation/infra/pkg/network"
-	"github.com/kaizen-experimentation/infra/pkg/observability"
-	"github.com/kaizen-experimentation/infra/pkg/secrets"
-	"github.com/kaizen-experimentation/infra/pkg/storage"
-	"github.com/kaizen-experimentation/infra/pkg/streaming"
-	"github.com/kaizen-experimentation/infra/pkg/waf"
+	"github.com/kaizen-experimentation/infra/pkg/aws/database"
+	"github.com/kaizen-experimentation/infra/pkg/aws/dns"
+	"github.com/kaizen-experimentation/infra/pkg/aws/loadbalancer"
+	"github.com/kaizen-experimentation/infra/pkg/aws/network"
+	"github.com/kaizen-experimentation/infra/pkg/aws/observability"
+	"github.com/kaizen-experimentation/infra/pkg/aws/secrets"
+	"github.com/kaizen-experimentation/infra/pkg/aws/storage"
+	"github.com/kaizen-experimentation/infra/pkg/aws/streaming"
+	"github.com/kaizen-experimentation/infra/pkg/aws/waf"
 )
 
 // Deploy is the main Pulumi program, exported for testability via the mock framework.
