+++ infra/pkg/aws/cicd/ecr_test.go
