+++ infra/pkg/aws/streaming/schema_registry.go
